requirements.txt
coreapi==2.3.3
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
